🎨 link 모듈을 common 모듈과 통합
독자적인 인터페이스를 노출하는 모듈이 아니고,
링크 컴포넌트에서 사용하는 코드이기 때문에 common으로 통합 관리합니다.
