+++ packages/router/src/common/router-guarded-link.tsx
